requirements.txt
Flask==1.1.4
google-generativeai==0.4.0
